relook into roxygen2 parsing again
It IS viable!

The caveat is that roxygen2 likes to evaluate the code that are after the blocks, so we just make the whole thing a roxygen comment. This can be done by looking for blocks that don't start with #' and adding it in.

After it gets parsed, we have the blocks separated out and can loop through them any way we please.