add helper functions to tests
I can now test knitting in a relative path. The only weird thing is that
source-ing seems to want to always default to the global environment and
I get errors like knitr_fig_path not found. To solve this, I had to add
`local = TRUE` to the source call, but it's definitely not ideal.