attempt to add user-specified params
I've run into a race-condition with caching. It turns out that evaluating cached knitr code inside a cached knitr chunk makes for a weird race condition where you have to run the document, let it error on the cached chunk, and run it again to get it to work. I've tried a lot of things, but nothing seems to be working.

Even explicitly setting cache to false does not help.